# .swagger-spec.json
{"openapi":"3.0.0","paths":{"/customer":{"post":{"operationId":"CustomerController_addCustomer","summary":"Adds a new Customer","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Customer"]},"get":{"operationId":"CustomerController_getAll","summary":"Gets all the Customer","parameters":[{"name":"searchKey","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerDto"}}}}}},"tags":["Customer"]}},"/customer/{id}":{"get":{"operationId":"CustomerController_getCustomer","summary":"Gets a Customer with id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerGetResponse"}}}}},"tags":["Customer"]},"put":{"operationId":"CustomerController_updateCustomer","summary":"Updates an Existing Customer","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerUpdateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Customer"]},"delete":{"operationId":"CustomerController_deleteCustomer","summary":"Delete an Existing Customer","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Customer"]}},"/transaction":{"post":{"operationId":"TransactionController_addTransaction","summary":"Adds a new Transaction","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Transactions"]},"get":{"operationId":"TransactionController_getAll","summary":"Gets all the Transaction","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionDto"}}}}}},"tags":["Transactions"]}},"/transaction/{id}":{"get":{"operationId":"TransactionController_getTransaction","summary":"Gets a Transaction with id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionGetResponse"}}}}},"tags":["Transactions"]},"put":{"operationId":"TransactionController_updateTransaction","summary":"Updates an Existing Transaction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionUpdateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Transactions"]},"delete":{"operationId":"TransactionController_deleteTransaction","summary":"Delete an Existing Transaction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Transactions"]}},"/moneyTransaction":{"post":{"operationId":"MoneyTransactionController_addMoneyTransaction","summary":"Adds a new MoneyTransaction","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoneyTransactionDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["MoneyTransaction"]},"get":{"operationId":"MoneyTransactionController_getAll","summary":"Gets all the MoneyTransaction","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoneyTransactionDto"}}}}}},"tags":["MoneyTransaction"]}},"/moneyTransaction/{id}":{"get":{"operationId":"MoneyTransactionController_getMoneyTransaction","summary":"Gets a MoneyTransaction with id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoneyTransactionGetResponse"}}}}},"tags":["MoneyTransaction"]},"put":{"operationId":"MoneyTransactionController_updateMoneyTransaction","summary":"Updates an Existing MoneyTransaction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoneyTransactionUpdateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["MoneyTransaction"]},"delete":{"operationId":"MoneyTransactionController_deleteMoneyTransaction","summary":"Delete an Existing MoneyTransaction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["MoneyTransaction"]}},"/admin/seed":{"post":{"operationId":"AdminController_seedAdmin","summary":"seed the first admin user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminDto"}}}},"responses":{"201":{"description":""}},"tags":["Admin"]}},"/admin/login":{"post":{"operationId":"AdminController_loginAdmin","summary":"login the admin user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Admin"]}}},"info":{"title":"PartyLedger","description":"A simple and Easy to Use Party Ledger","version":"1.0","contact":{}},"tags":[{"name":"Khata","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CustomerDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"contact":{"type":"string"}},"required":["name","email","address","contact"]},"ResponseDto":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"TransactionDto":{"type":"object","properties":{"customerId":{"type":"string"},"item":{"type":"string"},"unitPrice":{"type":"number"},"quantity":{"type":"number"},"totalPrice":{"type":"number"}},"required":["customerId","item","unitPrice","quantity","totalPrice"]},"CustomerGetResponse":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"contact":{"type":"number"},"transaction":{"$ref":"#/components/schemas/TransactionDto"}},"required":["name","email","address","contact","transaction"]},"CustomerUpdateDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"contact":{"type":"string"}},"required":["name","email","address","contact"]},"TransactionGetResponse":{"type":"object","properties":{"customerId":{"type":"string"},"item":{"type":"string"},"unitPrice":{"type":"number"},"quantity":{"type":"number"},"totalPrice":{"type":"number"}},"required":["customerId","item","unitPrice","quantity","totalPrice"]},"TransactionUpdateDto":{"type":"object","properties":{"customerId":{"type":"string"},"item":{"type":"string"},"unitPrice":{"type":"number"},"quantity":{"type":"number"},"totalPrice":{"type":"number"}},"required":["customerId","item","unitPrice","quantity","totalPrice"]},"MoneyTransactionDto":{"type":"object","properties":{"customerId":{"type":"string"},"amount":{"type":"number"}},"required":["customerId","amount"]},"MoneyTransactionGetResponse":{"type":"object","properties":{"customerName":{"type":"string"},"amount":{"type":"number"}},"required":["customerName","amount"]},"MoneyTransactionUpdateDto":{"type":"object","properties":{"amount":{"type":"number"}},"required":["amount"]},"AdminDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["name","email","password"]},"AdminLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}